Fix : message send NotificationService

+++ src/main/java/com/project/danim_be/notification/service/NotificationService.java
@@ -1,9 +1,12 @@
 package com.project.danim_be.notification.service;
 
+import com.project.danim_be.chat.entity.MemberChatRoom;
+import com.project.danim_be.chat.repository.MemberChatRoomRepository;
 import com.project.danim_be.common.exception.CustomException;
 import com.project.danim_be.common.exception.ErrorCode;
+import com.project.danim_be.member.entity.Member;
+import com.project.danim_be.member.repository.MemberRepository;
 import com.project.danim_be.notification.repository.EmitterRepository;
-import com.project.danim_be.notification.repository.NotificationRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -11,6 +14,7 @@ import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import java.io.IOException;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Slf4j
 @Service
@@ -18,8 +22,9 @@ import java.util.List;
 public class NotificationService {
     private final static Long DEFAULT_TIMEOUT = 3600000L;
     private final static String NOTIFICATION_NAME = "notify";
-    private final NotificationRepository notificationRepository;
     private final EmitterRepository emitterRepository;
+    private final MemberChatRoomRepository memberChatRoomRepository;
+    private final MemberRepository memberRepository;
 
     public SseEmitter connectNotification(Long userId) {
         // 새로운 SseEmitter를 만든다
@@ -48,15 +53,23 @@ public class NotificationService {
     public void send(List<Long> userIdList, Long messageId) {
         // 유저 ID로 SseEmitter를 찾아 이벤트를 발생 시킨다.
         for (Long userId : userIdList) {
-            emitterRepository.get(userId).ifPresentOrElse(sseEmitter -> {
-                try {
-                    sseEmitter.send(SseEmitter.event().id(messageId.toString()).name(NOTIFICATION_NAME).data("New notification"));
-                } catch (IOException exception) {
-                    // IOException이 발생하면 저장된 SseEmitter를 삭제하고 예외를 발생시킨다.
-                    emitterRepository.delete(userId);
-                    throw new CustomException(ErrorCode.FAIL_SEND_NOTIFICATION);
-                }
-            }, () -> log.info("No emitter found"));
+            Member member = memberRepository.findById(userId).orElseThrow(
+                    () -> new CustomException(ErrorCode.ID_NOT_FOUND)
+            );
+            MemberChatRoom memberChatRoom = memberChatRoomRepository.findByMember(member).orElseThrow(
+                    () -> new NoSuchElementException("없는 멤버챗룸임")
+            );
+            if (memberChatRoom.getRecentDisConnect().isBefore(memberChatRoom.getRecentConnect())) {
+                emitterRepository.get(userId).ifPresentOrElse(sseEmitter -> {
+                    try {
+                        sseEmitter.send(SseEmitter.event().id(messageId.toString()).name(NOTIFICATION_NAME).data("New notification"));
+                    } catch (IOException exception) {
+                        // IOException이 발생하면 저장된 SseEmitter를 삭제하고 예외를 발생시킨다.
+                        emitterRepository.delete(userId);
+                        throw new CustomException(ErrorCode.FAIL_SEND_NOTIFICATION);
+                    }
+                }, () -> log.info("No emitter found"));
+            }
         }
     }
 }
